requirements.txt
asgiref==3.8.1
boto3==1.35.90
botocore==1.35.90
dj-database-url==2.3.0
Django==5.1.4
django-heroku==0.3.1
django-storages==1.14.4
djangorestframework==3.15.2
gunicorn==23.0.0
jmespath==1.0.1
packaging==24.2
pillow==11.0.0
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
whitenoise==6.8.2
